Feat: Starting to move logging into routine
